+++ rf4s/utils.py
@@ -225,7 +225,7 @@ def reset_friction_brake_after(func):
             return
 
         with self.friction_brake_lock:
-            self.friction_brake.reset(self.cfg.BOT.FRICTION_BRAKE.INITIAL)
+            self.friction_brake.reset(self.cfg.FRICTION_BRAKE.INITIAL)
 
     return wrapper